# docs/design/api-spec.yaml
openapi: "3.0.3"
info:
  title: personTimeline API
  description: 多人物传记时间轴系统 — 第一阶段 API 规范
  version: "1.0.0"
  contact:
    name: personTimeline
servers:
  - url: http://localhost:8000
    description: 本地开发环境
paths:
...
  /api/upload:
    post:
      summary: 上传传记文件
      description: 上传 TXT 或 PDF 文件，提取文本后关联到人物
      operationId: uploadBiography
      tags: [上传与抽取]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, person_id]
              properties:
                file:
                  type: string
                  format: binary
                  description: 传记文件（.txt 或 .pdf）
                person_id:
                  type: string
                  format: uuid
                  description: 关联人物 ID
      responses:
        "201":
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BiographyText"
        "400":
          description: 不支持的文件格式
  /api/persons/{personId}/biography:
    get:
      summary: 获取人物传记文本列表
      operationId: listBiographyTexts
      tags: [上传与抽取]
      parameters:
        - $ref: "#/components/parameters/personId"
      responses:
        "200":
          description: 传记文本列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BiographyTextList"
  /api/biography/{biographyId}:
    delete:
      summary: 删除传记文本
      operationId: deleteBiography
      tags: [上传与抽取]
      parameters:
        - $ref: "#/components/parameters/biographyId"
      responses:
        "204":
          description: 删除成功（无内容）

  # ============================================================
  # AI 事件抽取
  # ============================================================
  /api/persons/{personId}/extract:
    post:
      summary: AI 事件抽取
      description: 调用大模型从传记文本中提取结构化事件并自动创建
      operationId: extractEvents
      tags: [上传与抽取]
      parameters:
        - $ref: "#/components/parameters/personId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractRequest"
      responses:
        "200":
          description: 抽取完成，返回创建的事件列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractResult"
        "504":
          description: AI 服务超时
components:
  parameters:
    personId:
      name: personId
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: 人物 ID
...
    biographyId:
      name: biographyId
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: 传记文本 ID
...
  schemas:
...
    TimeType:
      type: string
      enum: [POINT, PERIOD, FUZZY]
      description: 时间类型：时间点/时间段/模糊时段
    Granularity:
      type: string
      enum: [YEAR, MONTH, DAY, SEASON]
      description: 时间粒度
    EventType:
      type: string
      enum: [BIRTH, DEATH, EDUCATION, CAREER, CREATION, HISTORICAL, OTHER]
      description: 事件类型
...
    BiographyText:
      type: object
      required: [id, person_id, raw_text, created_at]
      properties:
        id:
          type: string
          format: uuid
        person_id:
          type: string
          format: uuid
        source_file:
          type: string
          nullable: true
        page:
          type: integer
          nullable: true
        text_length:
          type: integer
          description: 文本长度
        created_at:
          type: string
          format: date-time
    BiographyTextItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source_file:
          type: string
          nullable: true
        text_length:
          type: integer
        created_at:
          type: string
          format: date-time
    BiographyTextList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BiographyTextItem"

    # ---------- AI 抽取 ----------
    ExtractRequest:
      type: object
      properties:
        biography_id:
          type: string
          format: uuid
          nullable: true
          description: 指定抽取的文本 ID，为空则使用该人物所有文本
        model:
          type: string
          enum: [qwen, gpt, default]
          default: default
          description: 使用的大模型
    ExtractEventItem:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        display_time:
          type: string
          nullable: true
        time_type:
          $ref: "#/components/schemas/TimeType"
        granularity:
          $ref: "#/components/schemas/Granularity"
        event_type:
          $ref: "#/components/schemas/EventType"
        location:
          type: object
          nullable: true
        event_id:
          type: string
          format: uuid
          description: 已创建的事件 ID
        is_inferred:
          type: boolean
    ExtractResult:
      type: object
      properties:
        total:
          type: integer
          description: 抽取的事件总数
        events:
          type: array
          items:
            $ref: "#/components/schemas/ExtractEventItem"